% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{all_in_one}
\alias{all_in_one}
\title{all_in_one()}
\usage{
all_in_one(vector)
}
\arguments{
\item{vector}{is a list of entities}
}
\description{
function that uses functions "Scholar_is_it_a_method()" and "is_it_aTech()" to determine whether an entity is a technology, method or neither of them
}
