% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{is_it_aTech}
\alias{is_it_aTech}
\title{is_it_aTech()}
\usage{
is_it_aTech(entity)
}
\arguments{
\item{entity}{is a string and represents the entity a user wants to know if it is a technology or not}
}
\description{
function that predicts the class of a new entity through a pre trained KNN classifier
}
