% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarshMellow.R
\docType{package}
\name{MarshMellow}
\alias{MarshMellow}
\title{MarshMellow Package}
\description{
the package has built-in functions to scrape google scholar and that use a trained classifier (KNN) to determine whether a specific entity given in input by the user is a technology, a method or neither of them.
}
\author{
Pietro Manfredi \email{p.manfredi95@gmail.com}
}
