% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{Scholar_is_it_a_method}
\alias{Scholar_is_it_a_method}
\title{Scholar_is_it_a_method()}
\usage{
Scholar_is_it_a_method(entity)
}
\arguments{
\item{entity}{is given in input by the user and is a string}
}
\description{
function that web scrapes google scholar and determines whether a specific entity is a method or not
}
