% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{WikiCateg}
\alias{WikiCateg}
\title{WikiCateg()}
\usage{
WikiCateg(pagetitle)
}
\arguments{
\item{pagetitle}{is the title of the page given in input by the user and has to be a string}
}
\description{
function that extracts the categories of a Wikipedia page
}
